My VA - Remove deprecated v0/appointments code
