+++ src/applications/personalization/dashboard/mocks/feature-toggles/index.js
@@ -1,7 +1,6 @@
 const generateFeatureToggles = (toggles = {}) => {
   const {
     myVaUseExperimental = true,
-    profileUseVaosV2Api = true,
     showMyVADashboardV2 = true,
     showPaymentAndDebtSection = true,
   } = toggles;
@@ -13,10 +12,6 @@ const generateFeatureToggles = (toggles = {}) => {
         {
           name: 'my_va_experimental',
           value: myVaUseExperimental,
-        },
-        {
-          name: 'profile_use_vaos_v2_api',
-          value: profileUseVaosV2Api,
         },
         {
           name: 'show_myva_dashboard_2_0',
