+++ src/applications/personalization/dashboard/mocks/server.js
@@ -9,7 +9,7 @@ const { createHealthCareStatusSuccess } = require('./health-care');
 const { createUnreadMessagesSuccess } = require('./messaging');
 const notifications = require('./notifications');
 const { user81Copays } = require('./medical-copays');
-const { v0, v2 } = require('./appointments');
+const { v2 } = require('./appointments');
 
 // set to true to simulate a user with debts for /v0/debts endpoint
 const hasDebts = false;
@@ -18,7 +18,6 @@ const hasDebts = false;
 const responses = {
   'GET /v0/feature_toggles': generateFeatureToggles({
     myVaUseExperimental: true,
-    profileUseVaosV2Api: true,
     showMyVADashboardV2: true,
     showPaymentAndDebtSection: true,
   }),
@@ -76,16 +75,6 @@ const responses = {
         userPercentOfDisability: 40,
       },
     },
-  },
-  'GET /vaos/v0/appointments': (req, res) => {
-    const { query } = req;
-    const { type } = query;
-
-    if (type === 'va' || type === 'cc') {
-      const rv = v0.createAppointmentSuccess(type);
-      return res.status(200).json(rv);
-    }
-    return res.status(400).json({ bad: 'type' });
   },
   'GET /vaos/v2/appointments': (_req, res) => {
     const rv = v2.createAppointmentSuccess({ startsInDays: [31] });
