+++ src/applications/personalization/dashboard/tests/e2e/cerner-messaging/header-is-still-showing.cypress.spec.js
@@ -12,16 +12,6 @@ describe('MyVA Dashboard - Appointments - v2', () => {
     cy.intercept('GET', '/data/cms/vamc-ehr.json', vamcErc);
   });
 
-  it('Header still exists when cerner message exists on V1', () => {
-    cy.intercept(
-      '/v0/feature_toggles*',
-      generateFeatureToggles({ showMyVADashboardV2: false }),
-    );
-    cy.visit('my-va/');
-    cy.injectAxeThenAxeCheck();
-    cy.get('[data-testid="cerner-widget"] > .hydrated').should('exist');
-    cy.findByTestId('health-care-section-header').should('exist');
-  });
   it('Header still exists when cerner message exists on V2', () => {
     cy.intercept(
       '/v0/feature_toggles*',
