+++ src/applications/personalization/dashboard/tests/e2e/health-care-cerner.cypress.spec.js
@@ -1,5 +1,6 @@
 import enrollmentStatusEnrolled from '@@profile/tests/fixtures/enrollment-system/enrolled.json';
 import vamcErc from '../fixtures/vamc-ehr.json';
+import { generateFeatureToggles } from '../../mocks/feature-toggles';
 
 import {
   makeUserObject,
@@ -53,6 +54,10 @@ describe('MyVA Dashboard - Cerner Widget', () => {
         'GET',
         '/v0/health_care_applications/enrollment_status',
         enrollmentStatusEnrolled,
+      );
+      cy.intercept(
+        '/v0/feature_toggles*',
+        generateFeatureToggles({ showMyVADashboardV2: true }),
       );
       cy.intercept('GET', '/data/cms/vamc-ehr.json', vamcErc);
     });
